Nop bai tap Module 2 - ss13 - Lan 2

+++ quan_ly_san_pham/src/model/Product.java
@@ -3,7 +3,7 @@ package model;
 public class Product {
     private String id;
     private String name;
-    private float price;
+    private int price;
     private int quantity;
     private String description;
 
@@ -11,7 +11,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(String id, String name, float price, int quantity, String description) {
+    public Product(String id, String name, int price, int quantity, String description) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -35,11 +35,11 @@ public class Product {
         this.name = name;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(int price) {
         this.price = price;
     }
